requirements.txt
Django==2.2.6
djangorestframework==3.10.3
entrypoints==0.3
mccabe==0.6.1
psycopg2==2.8.3
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.2
six==1.10.0
sqlparse==0.3.0
